refactor(ingest): fold ast_walker_match.go back into ast_walker.go
The 28-line ast_walker_match.go held only the astMatch struct + its
Values/Context/CaptureOrigin methods — supporting types for ASTWalker.Query.
No human would split that into its own file; move it into ast_walker.go
under the "Internal types" section, just above astNode.
