+++ internal/ingest/ast_walker.go
@@ -209,6 +209,33 @@ func SelectWalker(db *sql.DB) (Walker, error) {
 
 // --- Internal types ---
 
+// astMatch is the Match returned by ASTWalker.Query. Values holds resolved
+// capture text, captureRanges records byte ranges for write-back, and ctx
+// carries the ASTRoot scoped to the matched node.
+type astMatch struct {
+	values        map[string]any
+	captureRanges map[string][2]int // capture name → [startByte, endByte]
+	ctx           ASTRoot
+	startByte     int
+	endByte       int
+}
+
+func (m *astMatch) Values() map[string]any { return m.values }
+func (m *astMatch) Context() any           { return m.ctx }
+
+// CaptureOrigin satisfies OriginProvider for write-back support.
+// Returns byte ranges for @scope (the outer matched node) and any named
+// captures whose byte ranges were recorded during Query.
+func (m *astMatch) CaptureOrigin(name string) (uint32, uint32, bool) {
+	if name == "scope" {
+		return uint32(m.startByte), uint32(m.endByte), true
+	}
+	if r, ok := m.captureRanges[name]; ok && r[0] < r[1] {
+		return uint32(r[0]), uint32(r[1]), true
+	}
+	return 0, 0, false
+}
+
 type astNode struct {
 	id        string
 	parentID  string
